feat: add updateSettings to panel context
Sometimes some panel settings aren't available until the panel is
rendered and some more data is gathered for it - e.g., because it's
being asynchronously fetched like the title in panels-pages. This
context method allows the developer to override a panel's title and
background at runtime.
The allowed keys are background and title.