feat: show context panel if possible while snapping
Knowing your context is one of the best things Panels introduces and
that makes it different to every other experience in which content is
shown in silos, i.e., out of context.
When we introduced snapping earlier on we broke that a little bit by
forcing the user to either click back or slide back in order to see the
panel they came from because they were always looking at the focus
panel.
This commit introduces the formalised concept of a context panel, which
is the panel that is before the focus panel.

BREAKING CHANGE