fix: resetting of positions due to resize
Up until now, we were listening to the window's resize event and making
our runtime adjust to the screensize to match mobile devices or not.
Truth is that probably nobody apart from developers resize their windows
that often and if they do it's mainly to try something else.
Resizing the runtime so far has only led to an unexpected and rather
annoying flash in which the runtime moves to reposition itself as you
resize. The current behaviour is not to react to resize anymore.