feat: load app from window.panelsJson
```
<!-- ... -->
<script src=/panels.js></script>
<script src=/app.js></script>
<script>
  !function() {
    'use strict'

    window.panelsJson = {}
    window.panelsJson[location.hostname] = {
      module: {
        name: 'async'
      }
    }
  }()
</script>
<!-- ... -->
```
The script above allows you preload an app so that it doesn't need to be
fetched after your page loaded resulting in a faster boot time of your
app when it's accessed directly. This doesn't remove the need for
`panels.json` you can still only use that approach, however this should
be more performant.